+++ packages/browser/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- Added new method `isAndroid` to correctly identify Android devices [[#2299](https://github.com/Shopify/quilt/pull/2299)]
 
 ## 3.0.0 - 2022-05-19